Put the validators inside the eval Docker images
