+++ evals/hello/Dockerfile
@@ -1,6 +1,11 @@
 FROM python:alpine
-WORKDIR /home/gradbench
-COPY evals/hello/hello.adroit .
-COPY evals/hello/main.py .
-ENTRYPOINT ["python3", "/home/gradbench/main.py"]
+
+RUN pip install pydantic
+
+COPY python /home/gradbench/python
+ENV PYTHONPATH=/home/gradbench/python
+
+RUN ln -s /home/gradbench/python/gradbench/evals/hello/validate.py /home/gradbench/validate
+
+ENTRYPOINT ["python", "/home/gradbench/python/gradbench/evals/hello/run.py"]
 LABEL org.opencontainers.image.source=https://github.com/gradbench/gradbench
