--- evals/hello/validate.py
@@ -1,28 +1,0 @@
-#!/usr/bin/env python3
-
-import argparse
-from pathlib import Path
-
-from gradbench.validate import assert_equal, validate_fixed
-
-
-def checker(name, input, a, b):
-    assert_equal(a, b)
-
-
-def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--golden", required=True)
-    parser.add_argument("--log", required=True)
-    args = parser.parse_args()
-
-    validate_fixed(
-        module="hello",
-        golden=Path(args.golden).read_text(),
-        log=Path(args.log).read_text(),
-        checker=checker,
-    )
-
-
-if __name__ == "__main__":
-    main()
